backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-os-backend"
version = "0.1.0"
description = "FastAPI backend for Aegis OS emergency intelligence"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16,<2.0",
  "fastapi>=0.135,<0.136",
  "google-cloud-storage>=3.1,<4.0",
  "cloud-sql-python-connector[psycopg]>=1.18,<2.0",
  "google-genai>=1.10,<2.0",
  "pydantic-settings>=2.8,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "python-dotenv>=1.0,<2.0",
  "python-multipart>=0.0.20,<0.1",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.35,<0.36"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<0.29",
  "pytest>=8.3,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.11,<0.12"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
